backend/api-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
redis==5.0.1
pika==1.3.2
clickhouse-driver==0.2.6
PyJWT==2.10.1
